+++ changelog.md
@@ -1,10 +1,13 @@
+# v0.1.4-beta.1
+- Added **Search Event** and sorted the events in an alphabetical order
+- Added **Scale Player** Event
+- Added **Reverse Player** Event
+- Attempted to fix the issue with the **Color Player** settings action crash on Android
 # v0.1.3-beta.1
 - Fixed/Adjusted the Action Node UI
 - Removed Unnecessary/useless data stored in the Commands JSON
 - Code Cleanup on the Command Settings
 - Fixed the crash when Android users attempted to open settings for **Color Player** 
-- Added **Scale Player** Event
-- Added **Reverse Player** Event
 # v0.1.2-beta.1
 - Fixed the **Edit Camera** event having a misleading label and fixed the default value
 - Fixed the issues with Commands not saving correctly after quitting the game on Android + iOS